Move scratch to it's own header and source files
