+++ include/hx/thread/Thread.hpp
@@ -10,26 +10,6 @@ namespace hx
 {
 	namespace thread
 	{
-		class Scratch final
-		{
-			using ReleaseFunc = void(*)(cpp::marshal::View<uint8_t>);
-
-			int* count;
-			ReleaseFunc release;
-
-		public:
-			static Scratch alloc(int bytes);
-
-			cpp::marshal::View<uint8_t> view;
-
-			Scratch(int* _count, cpp::marshal::View<uint8_t> _view, ReleaseFunc _release);
-			Scratch(const Scratch& other);
-
-			~Scratch();
-
-			Scratch& operator=(const Scratch& other);
-		};
-
 		struct Thread_obj : public hx::Object
 		{
 			static Thread create(Callable<void(void)> job);
