+++ toolchain/haxe-target.xml
@@ -80,6 +80,7 @@
   <depend name="${HXCPP}/include/hx/thread/CountingSemaphore.hpp"/>
   <depend name="${HXCPP}/include/hx/thread/Thread.hpp"/>
   <depend name="${HXCPP}/include/hx/thread/ThreadLocal.hpp"/>
+  <depend name="${HXCPP}/include/hx/thread/Scratch.hpp"/>
   <section unless="HXCPP_LINK_NO_TOOL_DEPENDS" >
      <depend name="${HXCPP}/toolchain/haxe-target.xml" dateOnly="true" />
      <depend name="${HXCPP}/toolchain/${toolchain}-toolchain.xml" ifExists="${HXCPP}/toolchain/${toolchain}-toolchain.xml" dateOnly="true" />
@@ -216,6 +217,7 @@
   <file name="src/hx/thread/ConditionVariable.cpp"/>
   <file name="src/hx/thread/ThreadImpl.cpp"/>
   <file name="src/hx/thread/ThreadLocal.cpp"/>
+  <file name="src/hx/thread/Scratch.cpp"/>
   
   <section if="windows">
     <file name="src/hx/thread/CountingSemaphore.win32.cpp"/>
